+++ pypaperless/endpoints.py
@@ -1,0 +1,1 @@
+
