+++ pypaperless/model/__init__.py
@@ -1,0 +1,9 @@
+"""PyPaperless model."""
+
+from .documents import Document
+from .helpers import DocumentHelper
+
+__all__ = (
+    "Document",
+    "DocumentHelper",
+)
