+++ pypaperless/model/documents.pyi
@@ -1,0 +1,12 @@
+from typing import Any
+
+from pypaperless import Paperless as Paperless
+
+from .base import PaperlessBase as PaperlessBase
+
+class Document(PaperlessBase):
+    def __call__(self, pk: int): ...
+    def __init__(self, api: Paperless, _data: dict[str, Any] | None = None) -> None: ...
+
+    id: int = ...
+    title: str = ...
